+++ HISTORY.md
@@ -1,3 +1,10 @@
+0.1.0 / 2014-05-07
+==================
+
+  * Implemented plugin system for storage.
+  * Moved JSON store support out into its own plugin
+    ([tiq-json](http://github.com/imiric/tiq-json)).
+
 0.0.1 / 2014-05-02
 ==================